requirements.txt
camelot-py==0.9.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
dnspython==2.6.1
et-xmlfile==1.1.0
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pdfminer.six==20231228
pillow==10.2.0
pycparser==2.21
pymongo==4.6.2
pyparsing==3.1.1
PyPDF2==2.12.1
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tzdata==2024.1
